Move store to model
